refactor(data): transformation by source
