+++ data-preparation/src/source_aggregations/footprint_vs_territorial.py
@@ -2,7 +2,7 @@
 Footprint versus territorial emissions
 """
 import pandas as pd
-from src.transformation.demographic.countries import \
+from src.source_aggregations.demographic.countries import \
     StatisticsPerCountriesAndZonesJoiner
 from src.utils.translation import CountryTranslatorFrenchToEnglish
