+++ data-preparation/src/source_aggregations/ghg/fao.py
@@ -1,5 +1,5 @@
 import pandas as pd
-from src.transformation.demographic.countries import \
+from src.source_aggregations.demographic.countries import \
     StatisticsPerCountriesAndZonesJoiner
 from src.utils.format import StatisticsDataframeFormatter
 from src.utils.translation import CountryTranslatorFrenchToEnglish
