+++ data-preparation/src/sources/eia/eia.py
@@ -3,7 +3,7 @@ import os
 
 import pandas as pd
 import requests
-from src.transformation.demographic.countries import \
+from src.source_aggregations.demographic.countries import \
     StatisticsPerCountriesAndZonesJoiner
 from src.utils.format import StatisticsDataframeFormatter
 from src.utils.iso3166 import countries
